refactor: util의 하위로 패키지 이동
