+++ app/src/main/java/com/into/websoso/common/util/tracker/Tracker.kt
@@ -1,4 +1,4 @@
-package com.into.websoso.data.tracker
+package com.into.websoso.common.util.tracker
 
 interface Tracker {
     fun trackEvent(
