+++ app/src/main/java/com/into/websoso/common/util/tracker/amplitude/AmplitudeTracker.kt
@@ -1,10 +1,10 @@
-package com.into.websoso.data.tracker.amplitude
+package com.into.websoso.common.util.tracker.amplitude
 
 import android.content.Context
 import com.amplitude.android.Amplitude
 import com.amplitude.android.Configuration
 import com.amplitude.android.autocaptureOptions
-import com.into.websoso.data.tracker.Tracker
+import com.into.websoso.common.util.tracker.Tracker
 
 class AmplitudeTracker(context: Context, apiKey: String) : Tracker {
     private val amplitude =
