+++ app/src/main/java/com/into/websoso/ui/createFeed/CreateFeedSearchNovelBottomSheetDialog.kt
@@ -12,7 +12,7 @@ import com.into.websoso.R.layout.dialog_create_feed_search_novel
 import com.into.websoso.common.ui.base.BaseBottomSheetDialog
 import com.into.websoso.common.util.InfiniteScrollListener
 import com.into.websoso.common.util.SingleEventHandler
-import com.into.websoso.data.tracker.Tracker
+import com.into.websoso.common.util.tracker.Tracker
 import com.into.websoso.databinding.DialogCreateFeedSearchNovelBinding
 import com.into.websoso.ui.createFeed.adapter.SearchNovelAdapter
 import com.into.websoso.ui.createFeed.adapter.SearchNovelItemType.Loading
